Structuring, Bugfix negative altitude, GPS Sanity check.

+++ OpenTXLog2GoogleEarth/Configuration.cs
@@ -14,7 +14,7 @@
         /// NODE (KNOTEN) = 1.852
         /// MPH = 1.609344
         /// </summary>
-        public decimal GPSSpeedFactor { get; set; } = 1m;
+        public double GPSSpeedFactor { get; set; } = 1d;
         /// <summary>
         /// Label for speed output.
         /// </summary>
